Improvements and cleanup for readability
a:link, a:hover, and a:active were not working becuase they have a certain order they must be in. Figured this out and improved behavior of page.  Changed .active class colors, and removed an id tag that was not being used in the index.html. cleaned up html formatting to make it more human readable.