Lesson 6 - Rough Draft
All Pages, Major Functionality, JavaScript, and CSS present. Still needs the rest of the images resized and configured to be responsive. min-width needs to be resolved in a way that keeps the pages without scrollbars on the bottom. Page weight testing needs to be done and page reviewed with Responsive Testing & Assesment.