Modularization, attribution, and Nav working
Improved Java Scripts added some hacks and work arounds to decided what pages  attribution should be shown on, even though it will always be in the footer, it will be turned off for pages that do not require it. Same with the Hero Banners for the Town Pages. They will only have the banner if it is one of the town pages, others will not have it displayed.