Topic 8 completed
originally written to trigger transition with onmouseenter and onmouseleave, but then realized that users with mobile devices and tablets will not have a mouse cursor. changed it to onclick and created a toggle method. The previous methods are still purposely present, to assist later as a refrence point if something similar needs to be done in a future project.